changed dbjoin windows (and thus the main js window class) can now accept percentage values

+++ plugins/fabrik_element/databasejoin/databasejoin.js
@@ -96,7 +96,7 @@ var FbDatabasejoin = new Class({
 		};
 		var winWidth = this.options.windowwidth;
 		if (winWidth !== '') {
-			this.windowopts.width = winWidth.toInt();
+			this.windowopts.width = winWidth;
 			this.windowopts.onContentLoaded = onContentLoaded;
 		}
 		
@@ -436,6 +436,7 @@ var FbDatabasejoin = new Class({
 		url += "&triggerElement=" + this.element.id;
 		url += "&resetfilters=1";
 		url += '&c=' + this.options.listRef;
+		
 		this.windowopts = {
 			'id': id,
 			'title': Joomla.JText._('PLG_ELEMENT_DBJOIN_SELECT'),
@@ -443,7 +444,7 @@ var FbDatabasejoin = new Class({
 			'loadMethod': 'xhr',
 			'evalScripts': true,
 			'contentURL': url,
-			'width': this.options.windowwidth.toInt(),
+			'width': this.options.windowwidth,
 			'height': 320,
 			'minimizable': false,
 			'collapsible': true,
